requirements.txt
amadeus
fastapi
pydantic
uvicorn
openai
python-dotenv
pinecone
mangum